# Backend/api/openapi/teams-notification-api.yaml
openapi: 3.0.3
info:
  title: Teams Notification API (Go implementation)
  version: 1.0.0
  description: |
    This specification documents the REST endpoints implemented in the Teams Notification Go service.
    Includes all available routes for internal and external APIs.

    ## API Structure
    - **External APIs** (`/api/v1`): Public APIs for external clients
    - **Internal APIs** (`/internal/v1`): Internal management APIs
    - **System APIs**: Health, metrics, and configuration endpoints

    ## Recent Updates (v1.0.0)
    - **Unified Configuration**: Environment-based configuration system
    - **Error Handling**: Standardized error responses with proper HTTP status codes
    - **Token Management**: Redis-backed token caching for improved performance
    - **Monitoring**: Enhanced health checks and system metrics
    - **Logging**: Structured logging with configurable levels and formats
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.teamsnotify.com
    description: Production server
paths:
...
  /internal/v1/destinations/bot/{botId}:
    get:
      summary: Get destinations by bot
      tags: [Internal API - Destinations]
      description: Get destinations by bot ID
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of destinations for bot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationListResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"
...
  /internal/v1/bots/platform:
    get:
      summary: List Teams bots
      tags: [Internal API - Bots]
      description: Get a list of Teams bot services
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
        - $ref: "#/components/parameters/SearchParam"
      responses:
        "200":
          description: List of Teams bots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotListResponse"

    post:
      summary: Create Teams bot
      tags: [Internal API - Bots]
      description: Create a new Teams bot service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotCreateRequest"
      responses:
        "201":
          description: Bot created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
...
  /internal/v1/bots/status/{status}:
    get:
      summary: Get bots by status
      tags: [Internal API - Bots]
      description: Get bots by status
      parameters:
        - name: status
          in: path
          required: true
          schema:
            type: string
            enum: [active, inactive, error]
      responses:
        "200":
          description: List of bots with status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotListResponse"

  # Billing endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    LimitParam:
      name: limit
      in: query
      description: Number of items to return
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    OffsetParam:
      name: offset
      in: query
      description: Number of items to skip
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    SearchParam:
      name: search
      in: query
      description: Search term
      required: false
      schema:
        type: string
...
  responses:
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFoundError:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Detailed error information
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
...
    PaginationResponse:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        pages:
          type: integer

    # Bot Framework schemas
...
    DestinationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [personal, group, channel]
        conversationId:
          type: string
        displayName:
          type: string
        tenantId:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time

    # Configuration schemas
...
    DestinationListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/DestinationResponse"
        pagination:
          $ref: "#/components/schemas/PaginationResponse"

    # Bot schemas
    BotCreateRequest:
      type: object
      required:
        - name
        - appId
        - appPassword
        - capabilities
      properties:
        name:
          type: string
          description: Bot name
        appId:
          type: string
          description: Microsoft App ID
        appPassword:
          type: string
          description: Microsoft App Password
        capabilities:
          type: array
          items:
            type: string
          description: Bot capabilities
        status:
          type: string
          enum: [active, inactive, error]
          default: active
        description:
          type: string
          description: Bot description
...
    BotResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        appId:
          type: string
        capabilities:
          type: array
          items:
            type: string
        status:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BotListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/BotResponse"
        pagination:
          $ref: "#/components/schemas/PaginationResponse"
...